+++ app_spec/zomes/blog/code/src/lib.rs
@@ -18,6 +18,7 @@ use hdk::{
     holochain_core_types::{cas::content::Address, entry::Entry, json::JsonString, error::HolochainError},
     holochain_wasm_utils::api_serialization::get_links::GetLinksResult,
 };
+use blog::Env;
 
 define_zome! {
     entries: [
@@ -30,6 +31,12 @@ define_zome! {
 
     functions: {
         main (Public) {
+            show_env: {
+                inputs: | |,
+                outputs: |sum: ZomeApiResult<Env>|,
+                handler: blog::handle_show_env
+            }
+
             check_sum: {
                 inputs: |num1: u32, num2: u32|,
                 outputs: |sum: ZomeApiResult<JsonString>|,
@@ -64,6 +71,12 @@ define_zome! {
                 inputs: | |,
                 outputs: |post_hashes: ZomeApiResult<GetLinksResult>|,
                 handler: blog::handle_my_posts
+            }
+
+            my_posts_immediate_timeout: {
+                inputs: | |,
+                outputs: |post_hashes: ZomeApiResult<GetLinksResult>|,
+                handler: blog::handle_my_posts_immediate_timeout
             }
 
             my_posts_as_committed: {